update notebook paths
