+++ impepdom/__init__.py
@@ -1,13 +1,10 @@
-from .mlp import MultilayerPerceptron
 from .dataset import PeptideDataset
 from .train_nn import train_nn
-from .experiment_runner import run_experiment, load_trained_model, plot_train_history
-from .example import X
+from .experiment_runner import run_experiment, plot_train_history
+from .store_manager import load_trained_model
 
 __all__ = [
-    'MultilayerPerceptron',
     'PeptideDataset',
-    'X',
     'train_nn',
     'run_experiment'
-    ]
+]
